buildkite insights: move generic search utility, minor refactorings
