+++ misc/python/materialize/buildkite_insights/annotation_search/search_result_presentation.py
@@ -11,74 +11,18 @@
 
 from typing import Any
 
-from materialize.buildkite_insights.annotation_search.search_utility import (
-    _search_value_to_pattern,
+from materialize.buildkite_insights.util.search_utility import (
+    highlight_match,
+    trim_match,
 )
 from materialize.terminal import (
     COLOR_CYAN,
-    COLOR_GREEN,
     STYLE_BOLD,
     with_formatting,
-    with_formattings,
 )
 
 SHORT_SEPARATOR = "----------"
 LONG_SEPARATOR = "-------------------------------------------------------------------------------------"
-
-
-def highlight_match(input: str, search_value: str, use_regex: bool) -> str:
-    case_insensitive_pattern = _search_value_to_pattern(search_value, use_regex)
-    match_replacement = with_formattings(r"\1", [COLOR_GREEN, STYLE_BOLD])
-    return case_insensitive_pattern.sub(match_replacement, input)
-
-
-def trim_match(input: str, search_value: str, use_regex: bool) -> str:
-    # We do not care about multiple occurrences within an annotation and focus on the first one.
-
-    input = input.strip()
-    case_insensitive_pattern = _search_value_to_pattern(search_value, use_regex)
-
-    max_chars_before_match = 300
-    max_chars_after_match = 300
-
-    match = case_insensitive_pattern.search(input)
-    assert match is not None
-
-    match_begin_index = match.start()
-    match_end_index = match.end()
-
-    # identify cut-off point before first match
-    if match_begin_index > max_chars_before_match:
-        cut_off_index_begin = input.find(
-            " ", match_begin_index - max_chars_before_match, match_begin_index
-        )
-
-        if cut_off_index_begin == -1:
-            cut_off_index_begin = match_begin_index - max_chars_before_match
-    else:
-        cut_off_index_begin = 0
-
-    # identify cut-off point after first match
-    if len(input) > match_end_index + 300:
-        cut_off_index_end = input.rfind(
-            " ", match_end_index, match_end_index + max_chars_after_match
-        )
-
-        if cut_off_index_end == -1:
-            cut_off_index_end = match_end_index + max_chars_after_match
-    else:
-        cut_off_index_end = len(input)
-
-    result = input[cut_off_index_begin:cut_off_index_end]
-    result = result.strip()
-
-    if cut_off_index_begin > 0:
-        result = f"[...] {result}"
-
-    if cut_off_index_end != len(input):
-        result = f"{result} [...]"
-
-    return result
 
 
 def print_before_search_results() -> None:
@@ -86,7 +30,7 @@ def print_before_search_results() -> None:
     print(LONG_SEPARATOR)
 
 
-def print_match(
+def print_annotation_match(
     build_number: str,
     build_pipeline: str,
     branch: str,
